requirements.txt
# Deep Learning and Computer Vision
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Image Processing
opencv-python>=4.6.0
pillow>=9.0.0
scikit-image>=0.19.0
albumentations>=1.3.0

# Face Detection and Analysis
face-alignment>=1.3.5
facexlib>=0.3.0

# Data Processing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0
tensorboard>=2.8.0
wandb>=0.12.0

# Backend API
fastapi>=0.85.0
uvicorn>=0.18.0
python-multipart>=0.0.5
aiofiles>=0.8.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
einops>=0.6.0
timm>=0.6.0

# Dataset Management
kaggle>=1.5.12
kagglehub>=0.2.0
gdown>=4.5.0

# Metrics and Evaluation
lpips>=0.1.4
pytorch-fid>=0.3.0

# For FFHQ dataset preprocessing
lmdb>=1.3.0

# Optional: For better performance
ninja>=1.10.0